I have separated the old method execute_ast_node as different files that now are static functions. The idea from now on is just to have a set of classes that let me implement the rest of the AST nodes.
The next big thing that I want to work on is within
the context of execution and the function/class system.

The rest of the AST nodes are going to be implemented
in the next commits.

As a side note, I have deleted all the changes with a
composer update and all this changes have been deleted
from the system. Thanks to vscode from saving me from
this mistake, almost 4 hours of work would have been lost.

Also, this is still a toy project, so I will not put myself
any presure on the architecture nor the development.

I'm enjoying the most the node recognition and the funny
and joy that gives me to study the internals of PHP and
doing on my way without not wanting to do anything perfect
from the ground.